requirements.txt
O365
ollama
python-dotenv